backend/pyproject.toml
[project]
name = "rtlphishletgenerator"
version = "1.0.0"
description = "Automated Evilginx Phishlet Generator for Red Team Engagements"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "playwright>=1.48.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "ruamel.yaml>=0.18.0",
    "litellm>=1.50.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "websockets>=13.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"
